Most promising approach so far, found a nasty weird bug in SimpleFlowPID of all places. Committing before breaking

+++ src/main/java/agents/firm/production/control/maximizer/algorithms/marginalMaximizers/CostEstimate.java
@@ -13,20 +13,20 @@ package agents.firm.production.control.maximizer.algorithms.marginalMaximizers;
 public class CostEstimate
 {
 
-    final private long marginalCost;
+    final private float marginalCost;
 
-    final private long totalCost;
+    final private float totalCost;
 
-    public CostEstimate(long marginalCost, long totalCost) {
+    public CostEstimate(float marginalCost, float totalCost) {
         this.marginalCost = marginalCost;
         this.totalCost = totalCost;
     }
 
-    final public long getMarginalCost() {
+    final public float getMarginalCost() {
         return marginalCost;
     }
 
-    final public long getTotalCost() {
+    final public float getTotalCost() {
         return totalCost;
     }
